My Study_Buddy Project
Study-buddy is an online platform for engineering student. This is a website which is helpful for engineering student to find internships, to be updated for college events like cultural as well as technical , competitions, hackathons , free or paid courses from different platform as per there requirements , previous year question paper and career guides.